HugChat-main/HugChat-main/requirements.txt
gradio>=4.0.0
huggingface_hub>=0.21.4
python-dotenv>=1.0.0
